+++ ChangeLog.md
@@ -10,6 +10,7 @@
 * Fixes bug in Text table output breaking Word and Html output (#126)
 * Fixes bug in call stack enumeration with nested functions containing `Section -Orientation` definitions (#121)
 * Adds Json plugin output (thanks @carceneaux)
+* Updates build pipeline with DotNet Core Azure Sign Tool
 
 ### 0.10.0 ###